separate ride modal pages into separate components, move ModalPageProps to Modal folder for shared use
