+++ frontend/src/components/Modal/types.ts
@@ -1,0 +1,6 @@
+import { ObjectType } from '../../types/index';
+
+export type ModalPageProps = {
+  onSubmit: (data: ObjectType) => void;
+  formData?: ObjectType;
+};
