+++ frontend/src/components/RideModal/Pages/index.ts
@@ -1,0 +1,3 @@
+export { default as RideTimesPage } from './RideTimes';
+export { default as DriverPage } from './Driver';
+export { default as RiderInfoPage } from './RiderInfo';
